tidy up field_setter proc macro PR
